+++ lib/functionals.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
-double area(long double** map, long double level);
+long double area(long double** map, long double level);
 
-double length(long double** map, long double level);
+long double length(long double** map, long double level);
