+++ lib/pml.hpp
@@ -2,4 +2,4 @@
 
 void pml_gen(unsigned int j, long double **array);
 
-void pml_gen(double theta, long double **array);
+void pml_gen(long double theta, long double **array);
